chore: use release-please to automate versioning, changelogs and GitHub releases
Using this action, as long as contributors follow the [conventional commit guidelines](https://www.conventionalcommits.org/en/v1.0.0/)
new versions, releases and changelog updates can be automated.
This action can also be extended to [automate npm publishing](https://github.com/googleapis/release-please-action#automating-publication-to-npm) on version-up.
For maintainers this reduces a lot of overhead in the "admin" of new
releases, and for consumers it increases clarity on new changes.
Note that for the PRs created by release-please to be able to trigger
further GitHub actions, a `token` will need to be provided. In this case
I've pointed to a secret called `RELEASE_PLEASE_TOKEN`. See
https://github.com/googleapis/release-please-action#github-credentials
for more information on this.